fix mobile nav overlay issue. fix various other layout issues
