+++ frontend/public/index.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html lang="en" class="h-full">
+<html lang="en">
   <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
@@ -7,8 +7,8 @@
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
     <title>Marci Forms</title>
   </head>
-  <body class="h-full">
+  <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root" class="h-full relative overflow-x-hidden"></div>
+    <div id="root"></div>
   </body>
 </html>
