+++ frontend/src/components/Dashboard/Forms/index.js
@@ -26,19 +26,19 @@ export default function DashboardForms() {
 
   return (
     <>
-      <ul className="grid grid-cols-1 gap-4 sm:grid-cols-2 p-16">
+      <ul className="grid grid-cols-1 gap-4 sm:grid-cols-2">
         {forms?.map((form) => (
           <div
             key={form.id}
-            className="bg-green-200 shadow sm:rounded-lg max-w-xl"
+            className="bg-green-200 shadow rounded-lg max-w-xl min-w-[300px]"
           >
-            <div className="sm:p-6 flex items-center justify-between">
+            <div className="p-6 flex items-center justify-between">
               <h3 className="text-base font-semibold leading-6 text-gray-900 text-ellipsis">
                 {form.title}
               </h3>
               <div className="">
-                <div className="sm:flex sm:items-center sm:justify-between">
-                  <div className="sm:mt-0 sm:flex-shrink-0 flex items-center gap-x-2">
+                <div className="flex items-center justify-between">
+                  <div className="mt-0 flex-shrink-0 flex items-center gap-x-2">
                     <Link
                       to={`/dashboard/forms/${form.id}/analytics`}
                       type="button"
@@ -54,7 +54,7 @@ export default function DashboardForms() {
                       <PencilSquareIcon className="w-6" />
                     </Link>
                   </div>
-                  <div className="mt-4 sm:ml-6 sm:mt-0 sm:flex-shrink-0">
+                  <div className="ml-6 flex-shrink-0">
                     <button
                       type="button"
                       className="inline-flex items-center rounded-md font-semibold text-red-900 ring-gray-300 hover:bg-red-300 p-2"
